chore: rename stores to state
