+++ packages/renderer/src/lib/chat/components/ExportButton.svelte
@@ -2,7 +2,7 @@
 import { type Chat } from '@ai-sdk/svelte';
 import { toast } from 'svelte-sonner';
 
-import { flowCreationData } from '/@/lib/flows/flow-creation-data.svelte';
+import { flowCreationData } from '/@/lib/chat/state/flow-creation-data.svelte';
 import { handleNavigation } from '/@/navigation';
 import { isFlowConnectionAvailable } from '/@/stores/flow-provider';
 import type { MCPRemoteServerInfo } from '/@api/mcp/mcp-server-info';
